issue 1588 - fixing date format
Trend was failing as date format extracted was incorrect.
Date format is being identified by picking up only first entry from the sample list and comparing with different formats listed in a format list. As the only one entry is used, getting exact format was an issue.
Changed the code to sort the date column list in descending order before sending it to extract the format so that the first value being picked is the largest.